Use context manager to close ThreadPool
On Pi the threads limit seems to be somewhere between 200 and 256, since the ThreadPool isn't closed these seem to hang around and cause RuntimeErrors when not enough threads can be spawned.

Using the context manager will close these explicityly (or at least implicitly (the point is that they get closed. Good)).