Install to and read skills from the XDG home directory
To be fully XDG compliant, we should be reading from the other XDG Data
directories as well (/usr/share and /usr/local/share), but that will
cause problems with skill_id's and the likes. For this reason we'll only
read from the user home directory for now ($XDG_DATA_HOME/mycroft/skills
or $HOME/mycroft/skills).

Also stop pip from ever using sudo. MSM will only install skills for a
user, not system-wide, and thus the Python dependencies should be
installed as such as well