Use git.Repo() to determine remote url
In newer GitPython the Git(path).config() seems to fall back to info
contained in the current directory leading to a possible mixup of origin
URLs. (as illustrated in the test case test_from_folder())

This replaces the check with a check using Repo.remote() to find the
remote for a repo at a location.